less terrible search resubmission
still room for improvement, including caching
